02_implementation/requirements.txt
# Core LangChain packages
langchain
langchain-core
langchain-community
langchain-ollama

# Tavily search integration
tavily-python


python-dotenv

